WellenceChallenge/pyproject.toml
[tool.poetry]
name = "wellencechallenge"
version = "0.1.0"
description = ""
authors = ["Rick Chang <rick800721@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-slick-reporting = "^1.3.1"
django-crispy-forms = "^2.3"
crispy-bootstrap4 = "^2024.1"
django-ninja = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
